// src/app/master/model/userrolemaster.ts
export class UserroleMaster {
    code: string="";
   empname:string = "";
   rolename: string ="";
   rolecode: string = "";
   username: string = "";
   password: string = "";
   status: boolean = false;

  
}
export class UserroleMasterEdit{
    rolename: string = "";
    rolecode: string = "";
    username: string = "";
    password: string = "";
    // code: string = "";
    // rolename: string = "";
    // empname: string = "";
    // password: string = "";
    // username:string="";
    status: boolean = false;
}
export class UserroleMasterAdd {
    employee: string = "";
    rolecode: string = "";
}  
   

export class UseremployeeSelected{
    employee: string = "";
    rolecode: string = "";
}
// }
// export class Usercodeselected{
//     rolecode: string = "";
// }